src: reclassify and rename some components
