+++ app/src/main/java/com/google/android/diskusage/ui/common/FileInfoAdapter.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package com.google.android.diskusage.ui.delete;
+package com.google.android.diskusage.ui.common;
 
 import java.io.File;
 import java.util.ArrayList;
