+++ app/src/main/res/layout/delete_view.xml
@@ -67,7 +67,7 @@
         android:layout_alignParentStart="true"
         android:layout_alignParentEnd="true"
         android:id="@+id/list" />
-    <com.google.android.diskusage.ui.MyTextView
+    <com.google.android.diskusage.ui.common.AdaptiveTextView
         android:id="@+id/summary"
         style="?android:attr/textAppearanceLarge"
         android:layout_width="match_parent"
